feat(website): new trending page and move cosmetics page
