+++ apps/website/src/routes/emotes/(directory)/(emotes)/trending/weekly/+page.ts
@@ -1,0 +1,5 @@
+import { redirect } from "@sveltejs/kit";
+
+export function load() {
+	redirect(301, "/emotes/trending");
+}
